requirements.txt
feedparser
python-telegram-bot==20.8
beautifulsoup4
pytz
deepl
requests